# backend/docs/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: CozyVTT API
  description: |
    RESTful API for CozyVTT — a self-hosted, browser-based Virtual Tabletop platform.

    ## Authentication
    All endpoints (except `/health`, `/api/setup/*`, and public auth routes) require authentication via session cookies.
    Login via `/api/auth/login` to obtain a session cookie.

    ## MFA Flow
    If a user has MFA enabled, `/api/auth/login` returns `{ mfaRequired: true }` instead of a full session.
    The client must then call `/api/auth/mfa/verify-login` with a TOTP code or backup code to complete authentication.

    ## Authorization
    - **Platform roles**: `ADMIN` (full system access), `USER` (campaign access)
    - **Campaign roles**: `DM` (full campaign control), `PLAYER` (limited interaction), `SPECTATOR` (read-only)

    ## Rate Limiting
    - Global API limit: 300 requests per minute per IP (all `/api/*` routes)
    - Authentication endpoints: 5 requests per 15 minutes
    - MFA setup verification: 5 attempts per 15 minutes
    - MFA login verification: 5 attempts per 15 minutes (separate limit)
    - Asset uploads: 30 requests per minute per user (configurable via `ASSET_UPLOAD_RATE_LIMIT`)
    - Chat messages: 5 per minute (WebSocket)
    - Dice rolls: 30 per minute (WebSocket)
    - Token movement: 60 per second (WebSocket, throttled)

    Rate-limited responses return HTTP 429 with `RateLimit-*` standard headers.

    ## Real-time Events
...

  version: 1.0.0
  contact:
    name: CozyVTT
...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/setup/init:
    post:
      tags:
        - Setup
      summary: Initialize system
      description: Create first admin account and complete setup wizard. Only works if setup is not yet complete.
      operationId: initializeSystem
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - displayName
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                displayName:
                  type: string
      responses:
        '201':
          description: Setup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================
  # AUTHENTICATION
  # ============================================
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account. First user automatically becomes ADMIN. Rate limited (5/15min).
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - displayName
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                displayName:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cozyvtt.sid
      description: Session cookie obtained via /api/auth/login
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        platformRole:
          type: string
          enum: [ADMIN, USER]
        mfaEnabled:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: BadRequest
          description: Short machine-readable error identifier
        message:
          type: string
          example: One or more fields failed validation.
          description: Human-readable error description
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional structured details (e.g. per-field Zod issues)
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadRequest
            message: Required field "name" is missing.
...
    RateLimited:
      description: |
        Too Many Requests - Rate limit exceeded. Response includes
        `RateLimit-Limit`, `RateLimit-Remaining`, and `RateLimit-Reset` headers.
      headers:
        RateLimit-Limit:
          schema: { type: integer }
          description: Maximum requests permitted in the current window
        RateLimit-Remaining:
          schema: { type: integer }
          description: Requests remaining in the current window
        RateLimit-Reset:
          schema: { type: integer }
          description: Seconds until the window resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: TooManyRequests
            message: Upload rate limit reached. Please slow down and try again in a minute.
...